Updated the README, renamed lambda to handlers, and added sql and model folders.
