+++ index.js
@@ -5,25 +5,18 @@ const Alexa = require('alexa-sdk');
 const pg = require("pg");
 
 //imports
-const config = require('./lambda/config');
-const mainHandlers = require('./lambda/main');
-const checkHandlers = require("./lambda/checkInOut");
-const moodHandlers = require("./lambda/moods");
-const statusHandlers = require("./lambda/status");
-const creatingHandlers = require("./lambda/createRole");
-const careHandlers = require("./lambda/createCare");
+const config = require('./handlers/config');
+const mainHandlers = require('./handlers/main');
+const checkHandlers = require("./handlers/checkInOut");
+const moodHandlers = require("./handlers/moods");
+const statusHandlers = require("./handlers/status");
+const creatingHandlers = require("./handlers/createRole");
+const careHandlers = require("./handlers/createCare");
 
 //handler
 exports.handler = function(event, context, callback) {
     const alexa = Alexa.handler(event, context, callback);
     alexa.appId = config.appId;
-    const pool = new pg.Pool({
-        user: config.dbUSER,
-        password: config.dbPWD,
-        database: config.dbDatabase,
-        host: config.dbURL,
-        port: config.dbPort
-    });
     context.callbackWaitsForEmptyEventLoop = false;
     alexa.registerHandlers(mainHandlers, checkHandlers, moodHandlers, statusHandlers, creatingHandlers, careHandlers);
     alexa.execute();
